Adjust seeder and some tables

+++ db/migrate/20200629010045_create_receive_order_details.rb
@@ -3,9 +3,9 @@ class CreateReceiveOrderDetails < ActiveRecord::Migration[6.0]
     create_table :receive_order_details do |t|
       t.belongs_to :receive_order, index: true, foreign_key: true
       t.belongs_to :item, index: true, foreign_key: true
-      t.numeric :cogs, null: false
-      t.numeric :price, null: false
-      t.numeric :quantity, null: false
+      t.integer :cogs, null: false
+      t.integer :price, null: false
+      t.integer :quantity, null: false
 
       t.timestamps
     end
