+++ db/migrate/20200629010514_create_cart.rb
@@ -1,7 +1,7 @@
 class CreateCart < ActiveRecord::Migration[6.0]
   def change
     create_table :carts do |t|
-      t.numeric :discount
+      t.integer :discount
 
       t.timestamps
     end
