+++ db/migrate/20200629011102_create_transaction_payments.rb
@@ -1,7 +1,7 @@
 class CreateTransactionPayments < ActiveRecord::Migration[6.0]
   def change
     create_table :transaction_payments do |t|
-      t.numeric :payment_sum, null: false
+      t.integer :payment_sum, null: false
 
       t.timestamps
     end
